dev:externalize GerritTriggerDescriptor
